+++ src/main/java/pl/asie/computronics/cc/IntegrationComputerCraft.java
@@ -11,6 +11,7 @@ import pl.asie.computronics.integration.appeng.DriverSpatialIOPort;
 import pl.asie.computronics.integration.buildcraft.DriverHeatable;
 import pl.asie.computronics.integration.cofh.DriverEnergyProvider;
 import pl.asie.computronics.integration.cofh.DriverEnergyReceiver;
+import pl.asie.computronics.integration.draconicevolution.DriverExtendedRFStorage;
 import pl.asie.computronics.integration.enderio.DriverAbstractMachine;
 import pl.asie.computronics.integration.enderio.DriverAbstractPoweredMachine;
 import pl.asie.computronics.integration.enderio.DriverCapacitorBank;
@@ -128,6 +129,11 @@ public class IntegrationComputerCraft {
 				registerMultiPeripheralProvider(new DriverCapacitorBankOld.CCDriver());
 				registerMultiPeripheralProvider(new DriverTransceiver.CCDriver());
 			}
+		}
+
+		if(Loader.isModLoaded(Mods.API.DraconicEvolution)
+			&& compat.isCompatEnabled(Compat.DraconicEvolution)) {
+			registerMultiPeripheralProvider(new DriverExtendedRFStorage.CCDriver());
 		}
 
 		if(Mods.API.hasVersion(Mods.API.BuildCraftTiles, "[1.1,)")) {